+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:nuliga_app/bottom_navigation.dart';
 import 'package:nuliga_app/services/followed-teams/followed_teams_provider.dart';
 import 'package:nuliga_app/services/localization/localization_provider.dart';
 import 'package:provider/provider.dart';
-import 'l10n/app_localizations.dart';
+import 'localization/app_localizations.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
