+++ lib/pages/dashboard/dashboard_section.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/pages/shared/constants.dart';
 
 class DashboardSection extends StatelessWidget {
