+++ lib/pages/dashboard/last-matches/last_matches.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/services/followed-teams/followed_club.dart';
 import 'package:nuliga_app/pages/dashboard/last-matches/last_matches_card.dart';
 import 'package:nuliga_app/services/matches/last_matches_service.dart';
@@ -14,7 +14,7 @@ class LastMatches extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context)!;
-    
+
     return SizedBox(
       height: 290,
       child: FutureBuilder(
