+++ lib/pages/next-match/next_match_countdown.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/pages/shared/constants.dart';
 
 class NextMatchCountdown extends StatefulWidget {
