+++ lib/pages/next-match/next_match_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/services/matches/next-matches/model/future_match.dart';
 import 'package:nuliga_app/pages/next-match/next_match_countdown.dart';
 import 'package:nuliga_app/pages/next-match/next_match_date_info.dart';
@@ -19,7 +19,7 @@ class NextMatchPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context)!;
-    
+
     return Scaffold(
       appBar: AppBar(title: Text(l10n.upcoming)),
       body: Padding(
