+++ lib/pages/settings/club_edit_dialog_step_team_name.dart
@@ -1,7 +1,7 @@
 import 'dart:developer' as developer;
 
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/services/league-table/model/league_team_ranking.dart';
 import 'package:nuliga_app/pages/shared/loading_indicator.dart';
 import 'package:nuliga_app/services/settings/settings_service.dart';
@@ -53,7 +53,7 @@ class _ClubEditDialogStepTeamNameState
   @override
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context)!;
-    
+
     return widget.rankingUrl.isEmpty
         ? Text(l10n.pleaseEnterLeagueUrlInStep1)
         : FutureBuilder<List<LeagueTeamRanking>>(
