+++ lib/pages/settings/localization_dropdown_menu.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/services/localization/localization_provider.dart';
 import 'package:provider/provider.dart';
