+++ lib/pages/shared/nothing_to_display_indicator.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 import 'package:nuliga_app/pages/settings/settings_page.dart';
 
 // TODO: use again
