+++ lib/services/localization/localization_provider.dart
@@ -1,7 +1,7 @@
 import 'dart:developer' as developer;
 
 import 'package:flutter/material.dart';
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 
 class LocalizationProvider extends ChangeNotifier {
   Locale _locale = const Locale("de");
