+++ lib/services/match-result/model/game_type.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: non_constant_identifier_names
 
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 
 enum GameType {
   ms1(parsingText: "1.HE"),
