+++ lib/services/matches/last-matches/model/match_result.dart
@@ -1,4 +1,4 @@
-import 'package:nuliga_app/l10n/app_localizations.dart';
+import 'package:nuliga_app/localization/app_localizations.dart';
 
 class MatchResult {
   final DateTime time;
